refine UI CC
