+++ src/components/handshakes/exchange/Feed/FeedCreditCard_BackUp.scss
@@ -1,0 +1,8 @@
+.buy-button {
+  background: #111111;
+  border-radius: 6px;
+  font-size: 20px;
+  color: #FFFFFF;
+  letter-spacing: 0;
+  text-align: center;
+}
